Removed validate from blob
