+++ src/main/kotlin/org/ptss/support/core/services/BlobStorageService.kt
@@ -7,10 +7,12 @@ import org.ptss.support.infrastructure.repositories.BlobStorageRepository
 import org.ptss.support.infrastructure.util.executeWithExceptionLoggingAsync
 import org.slf4j.LoggerFactory
 import java.io.InputStream
-import java.util.*
 
 @ApplicationScoped
-class BlobStorageService(private val blobStorageRepository: BlobStorageRepository) {
+class BlobStorageService(
+    private val blobStorageRepository: BlobStorageRepository,
+    private val generalInformationService: GeneralInformationService
+) {
 
     private val logger = LoggerFactory.getLogger(BlobStorageService::class.java)
 
@@ -19,8 +21,8 @@ class BlobStorageService(private val blobStorageRepository: BlobStorageRepositor
             val bufferedStream = stream.buffered()
             logger.executeWithExceptionLoggingAsync(
                 operation = {
-                    val fileType = detectFileTypeMagicNumbers(bufferedStream)
-                    val fileName = generateFileName(fileType)
+                    val fileType = generalInformationService.detectFileTypeMagicNumbers(bufferedStream)
+                    val fileName = generalInformationService.generateFileName(fileType)
                     blobStorageRepository.uploadFile(fileName, bufferedStream)
                 },
                 logMessage = "Error uploading file to Azure Blob Storage",
@@ -45,32 +47,5 @@ class BlobStorageService(private val blobStorageRepository: BlobStorageRepositor
                 )
             }
         )
-    }
-
-    private fun detectFileTypeMagicNumbers(fileStream: InputStream): String {
-        val buffer = ByteArray(8)
-        fileStream.mark(8)
-        fileStream.read(buffer)
-        fileStream.reset()
-
-        return when {
-            buffer.startsWith(byteArrayOf(0xFF.toByte(), 0xD8.toByte(), 0xFF.toByte())) -> ".jpg"
-            buffer.startsWith(byteArrayOf(0x89.toByte(), 0x50.toByte(), 0x4E.toByte(), 0x47.toByte())) -> ".png"
-            buffer.startsWith("RIFF".toByteArray()) && buffer.slice(8..11).toByteArray().contentEquals("WEBP".toByteArray()) -> ".webp"
-            buffer.startsWith(byteArrayOf(0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x18.toByte(), 0x66.toByte(), 0x74.toByte(), 0x79.toByte(), 0x70.toByte())) -> ".mp4"
-            buffer.startsWith(byteArrayOf(0x1A.toByte(), 0x45.toByte(), 0xDF.toByte(), 0xA3.toByte())) -> ".mkv"
-            buffer.startsWith("%PDF".toByteArray()) -> ".pdf"
-            else -> throw APIException(
-                errorCode = ErrorCode.MEDIA_CREATION_ERROR,
-                message = "Unsupported file type"
-            )
-        }
-    }
-
-    private fun ByteArray.startsWith(prefix: ByteArray): Boolean =
-        this.take(prefix.size).toByteArray().contentEquals(prefix)
-
-    private fun generateFileName(fileType: String): String {
-        return "${UUID.randomUUID()}$fileType"
     }
 }
